refactor ErrorCorector
